requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.1
python-jose==3.3.0
emails==0.6
jinja2==3.1.2
pymysql==1.1.0
